chore(route-recognizer): Tighten up types
Particularly, the new getRoute method was returning an any type because its type was not yet explicitly defined. Here I've explicitly defined the type, used it in the method, and updated all other types that were exposed as sloppy or wrong as a result. I've also added a new generic Segment type as an alias for any type of Segment, since they are often consolidated into a list and used similarly. It may make sense to create a proper interface that describes a generic segment in the future.